Hamamatsu 0620 2

+++ Project_K/Util/SelectUI.cpp
@@ -419,7 +419,7 @@ void SelectUI::DrawChar()
 
 	// 1Pが選択してるキャラ
 	my::MyDrawRectRotaGraph(
-		400, m_player1.m_posY,//プレイヤーの位置
+		400, static_cast<int>(m_player1.m_posY),//プレイヤーの位置
 		m_player1.m_imgWidth * m_indexWidth1, m_player1.m_imgHight * m_indexHight1,// 画像の右上
 		m_player1.m_imgWidth, m_player1.m_imgHight,// 画像の切り取り
 		m_player1.m_imgSize,// 画像の拡大率
@@ -430,7 +430,7 @@ void SelectUI::DrawChar()
 
 	// 2Pが選択してるキャラ
 	my::MyDrawRectRotaGraph(
-		Game::kScreenWidth - 400, m_player2.m_posY,//プレイヤーの位置
+		Game::kScreenWidth - 400, static_cast<int>(m_player2.m_posY),//プレイヤーの位置
 		m_player2.m_imgWidth * m_indexWidth2, m_player2.m_imgHight * m_indexHight2,// 画像の右上
 		m_player2.m_imgWidth, m_player2.m_imgHight,// 画像の切り取り
 		m_player2.m_imgSize,// 画像の拡大率
